pyproject.toml
# pyproject.toml
[project]
name = "git-to-blog"
version = "0.1.0"
description = "Automatically generate blog posts from GitHub commit activity."
authors = [
    { name = "Rajan Shrestha", email = "shrestharajan726@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "fastapi==0.115.12",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.25",
    "langchain-core==0.3.65",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.45",
    "orjson==3.10.18",
    "packaging==24.2",
    "pluggy==1.6.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pytest==8.4.0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "tenacity==9.1.2",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.3",
    "zstandard==0.23.0",
    "asyncpg==0.30.0"
]
